+++ src/Glyph.ts
@@ -1,0 +1,9 @@
+import * as pdi from "@akashic/pdi-types";
+import { SurfaceAtlas } from ".";
+
+export interface Glyph extends pdi.Glyph {
+	/**
+	 * @ignore
+	 */
+	_atlas: SurfaceAtlas | null;
+}