+++ src/SurfaceAtlasSet.ts
@@ -1,4 +1,4 @@
-import { CommonSize, Glyph, ResourceFactory, SurfaceAtlas as PdiSurfaceAtlas } from "@akashic/pdi-types";
+import { CommonSize, Glyph, ResourceFactory } from "@akashic/pdi-types";
 import { SurfaceAtlas } from "./SurfaceAtlas";
 
 function calcAtlasSize(hint: SurfaceAtlasSetHint): CommonSize {
@@ -49,7 +49,7 @@ export interface RemoveAtlasData {
 	/**
 	 * 削除対象のSurfaceAtlas
 	 */
-	surfaceAtlases: PdiSurfaceAtlas[];
+	surfaceAtlases: SurfaceAtlas[];
 
 	/**
 	 * 削除対象のグリフ
@@ -86,7 +86,7 @@ export class SurfaceAtlasSet {
 	/**
 	 * @private
 	 */
-	_surfaceAtlases: PdiSurfaceAtlas[];
+	_surfaceAtlases: SurfaceAtlas[];
 
 	/**
 	 * @private
@@ -172,7 +172,7 @@ export class SurfaceAtlasSet {
 	 * glyphが持つ情報をSurfaceAtlasへ移動し、移動したSurfaceAtlasの情報でglyphを置き換える。
 	 * @private
 	 */
-	_moveGlyphSurface(glyph: Glyph): PdiSurfaceAtlas | null {
+	_moveGlyphSurface(glyph: Glyph): SurfaceAtlas | null {
 		for (let i = 0; i < this._surfaceAtlases.length; ++i) {
 			const index = (this._currentAtlasIndex + i) % this._surfaceAtlases.length;
 			const atlas = this._surfaceAtlases[index];
@@ -241,7 +241,7 @@ export class SurfaceAtlasSet {
 	 * 通常、ゲーム開発者がこのメソッドを呼び出す必要はない。
 	 * @param index 取得対象のインデックス
 	 */
-	getAtlas(index: number): PdiSurfaceAtlas {
+	getAtlas(index: number): SurfaceAtlas {
 		return this._surfaceAtlases[index];
 	}
 
@@ -294,7 +294,7 @@ export class SurfaceAtlasSet {
 	 * 通常、ゲーム開発者がこのメソッドを呼び出す必要はない。
 	 * @param glyph グリフ
 	 */
-	addGlyph(glyph: Glyph): PdiSurfaceAtlas | null {
+	addGlyph(glyph: Glyph): SurfaceAtlas | null {
 		// グリフがアトラスより大きいとき、`_atlasSet.addGlyph()`は失敗する。
 		// `_reallocateAtlas()`でアトラス増やしてもこれは解決できない。
 		// 無駄な空き領域探索とアトラスの再確保を避けるためにここでリターンする。