+++ src/SurfaceAtlasSlot.ts
@@ -1,17 +1,15 @@
-import * as pdi from "@akashic/pdi-types";
-
 /**
  * SurfaceAtlasの空き領域管理クラス。
  *
  * 本クラスのインスタンスをゲーム開発者が直接生成することはなく、ゲーム開発者が利用する必要もない。
  */
-export class SurfaceAtlasSlot implements pdi.SurfaceAtlasSlot {
+export class SurfaceAtlasSlot {
 	x: number;
 	y: number;
 	width: number;
 	height: number;
-	prev: pdi.SurfaceAtlasSlot | null;
-	next: pdi.SurfaceAtlasSlot | null;
+	prev: SurfaceAtlasSlot | null;
+	next: SurfaceAtlasSlot | null;
 
 	constructor(x: number, y: number, width: number, height: number) {
 		this.x = x;