+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 3.0.0-beta.28
 
 その他変更
- * `@akashic/pdi-common-imple` の `SurfaceAtlas`, `SurfaceAtlasSlot` を Akashic Engine で持つよう変更
+ * @akashic/pdi-types@1.0.1 に追従
+    * `@akashic/pdi-common-impl` の `SurfaceAtlas`, `SurfaceAtlasSlot` を Akashic Engine で持つよう変更
 
 ## 3.0.0-beta.27